refactor(tui): extract toolbar tip constants to tui/constant/tips.ts
